requirements.txt
instagrapi==2.0.0
crewai==0.1.7
python-dotenv==1.0.0
pydantic==2.0.0
wget==3.2
openai==1.0.0